AdminThemes: rewrite theme import.
This was totally broken, see e.g. issue #655. For example, trying
to register a theme already in themes/ in the database ("Install
this theme" menu item) did nothing but deleting this theme.

Now it should work properly for unpacked theme packages (themes
imported by a ZIP) as well as for themes already in themes/.