Installer: do a real theme installation.
This way we get the real thing. It'll allow to remove all the
handcrafted SQL and XML installer code which just duplicates
the theme configuration (and was usually not up to date).

Two modules had to get fixed to not insist on a theme being
installed (in the database) at their installation time.

This also messed up hooks, image types and some other stuff, which
will get rectified in the following commit(s).